+++ canaille/__init__.py
@@ -44,6 +44,7 @@ def create_app(
     sentry_sdk = setup_sentry(app)
     try:
         setup_logging(app)
+
         backend = setup_backend(app, backend, init_backend)
         setup_security(app)
         setup_flask_converters(app)